changed from double to float

+++ db/migrate/20160512193704_add_price_to_listings.rb
@@ -1,5 +1,5 @@
 class AddPriceToListings < ActiveRecord::Migration
   def change
-    add_column :listings, :price, :double
+    add_column :listings, :price, :float
   end
 end
